fix error when writing file with empty content
Previously, attempting to write a local file with content of "" would set the resource Id to "", since the Id is based on a SHA1 of the content. This caused the file to be deleted during the apply step.

The resource Id of an empty file is now the SHA1 of the filename concatenated with the SHA1 of the content, so it will never be "".